Adds Basic Authentication to the administration interface. It must be forced as it will not be asked for. This is not I repeat not Basic Authentication for the cal.php etc interfaces it will only work with /admin/
Also added is a hidden div tag that contains the users id in a tag that
can be id'ed by username_"There Username" this allows you to grab their
ID so that you can add calendars address books etc programatically.

I will confess this is by no way the best way to programatically
manipulate the users but it was the best way that I could do it with
the time I had. A user visiting via the web page will see nothing
different from before.