Fix pdo install bug and sqlite permission requirement
This fixes the issue, that baikal requires write permissions on sqlite
folder and file even if you want to use Mysql.
It also fixes a bug, where a pdo fatal error occurs at installation,
when no sqlite driver is installed.
This reverts to the try clause in initMysql(), cause this gives a
valuable error message, if a user messes up his config file and mysql
fails.