Update README.md with link to french tutorial
Hello there :wave:,

I have created a guide in French for the installation and use of Baïkal. It is my way to thank you for this great tool. I made it from the German guide, although it is not a translation because it is different.

PS : This is my first contribution in a GitHub project so please be merciful, I don't know if I'm doing it the right way :innocent: